% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate.R
\name{Aggregate}
\alias{Aggregate}
\title{Aggregate}
\usage{
Aggregate(data = NA, group = NA, val = NA, fun = "mean", subsets = NA)
}
\arguments{
\item{data}{is a data.frame object that need to be processed}

\item{group}{are the dimension variables}

\item{val}{are the variables that needed calculated}

\item{subsets}{are subset of data object}
}
\value{
same as aggregate function
}
\description{
this is a wrapp version of aggregate function in R base package.
}
\examples{
Aggregate(iris,group='Species',val=c('Sepal.Length','Sepal.Width'),fun='mean')
Aggregate(iris,group='Species',val=c('Sepal.Length','Sepal.Width'),fun='function(x)median(x,na.rm=T)')
Aggregate(iris,group='Species',val=c('Sepal.Length','Sepal.Width'),
fun='mean',subsets=list(val=c('Species','Sepal.Length'),legal=list(c('setosa','virginica'),
c('[0,10]')),method=c('elements','ranges'),mode=c('character','numeric')))


}
\author{
sontron
}
