% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply.R
\name{Apply}
\alias{Apply}
\title{Apply
a little change to apply. avoid error for apply function to a data.frame that only contain one var.}
\usage{
Apply(data, vars, MARGIN, FUN, na.rm = T, ...)
}
\description{
Apply
a little change to apply. avoid error for apply function to a data.frame that only contain one var.
}
\author{
sontron
}
