% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biVars.R
\name{biVars}
\alias{biVars}
\title{biVars}
\usage{
biVars(data, xvars, yvars = "", alter = c("two.sided", "less",
  "greater")[1], paired = FALSE, confLevel = 0.95, nullHyp = 0,
  normalSampleSize = 100)
}
\description{
is a function of basic hTest collection based on data.frame
}
