% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descTab.R
\name{descTab}
\alias{descTab}
\title{descTab}
\usage{
descTab(Formula, data)
}
\description{
this function is a revised version of mytable in moonBook package.
mytable(moonBook) can deal no more than two group variables. in descTab function, when there are more than two grouping 
variables, they are pasted together as one single variable.(use table1 instead!)
}
\examples{
descTab(Formula='vs+gear+am~disp+mpg',data=mtcars)

}
