% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faS.R
\name{faS}
\alias{faS}
\title{faS}
\usage{
faS(data, vars, nfcts = 2, Rotate = c("none", "varimax", "quartimax",
  "bentlerT", "varmin", "equamax", "geominT", "bifactor", "promax",
  "oblimin", "simplimax", "bentlerQ", "geominQ", "biquartimin",
  "cluster")[1], Scores = c("regression", "Thurstone", "tenBerge",
  "Anderson", "Barlett")[1], FM = c("minres", "uls", "ols", "wls", "gls",
  "pa", "ml", "minchi", "minrank")[1], subset = "all", faVarName = "",
  addVar = T)
}
\description{
factor analysis used in shiny app.
}
