% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplt.R
\name{ggplt2S}
\alias{ggplt2S}
\title{ggplt2S}
\usage{
ggplt2S(data, x, y = "NULL", size = "NULL", fill = "NULL",
  color = "NULL", shape = "NULL", alpha = "NULL", facetVar = "NULL",
  geom = c("box", "hist", "bar", "line", "jitter", "point", "smooth")[1],
  labx = "x", laby = "y", title = "my Plot", theme = c("grey", "bw",
  "classic", "dark")[1], smoothMethod = c("lm", "glm", "loess", "gam")[1],
  barPos = c("stack", "dodge")[1], Bins = "NULL", Colour = "NULL",
  Fill = "NULL", Size = "NULL", Alpha = "NULL", Width = "NULL",
  Shape = "NULL")
}
\arguments{
\item{data}{data.frame}

\item{x}{x vars.}

\item{y}{y vars, could be NULL}

\item{size}{variable mapping to size aes.}

\item{fill}{variable mapping to fill aes.}

\item{color}{variable mapping to color aes.}

\item{shape}{variable mapping to shape aes.}

\item{alpha}{variable mapping to alpha aes.}

\item{facetVar}{variables used in faceting.}

\item{geom}{geoms for ploting, eg: geom=c('point','smooth')}

\item{labx}{x lable.}

\item{laby}{y lable.}

\item{title}{plot title.}

\item{theme}{could be "grey","bw","classic","dark",etc.}

\item{smoothMethod}{smoothing method used in a smooth line.}

\item{barPos}{used when geom including "bar",could be "stack","dodge".}

\item{Bins}{numeric value to set binwidth in a histogram plot.}

\item{Colour}{setting colour, eg:Colour='red'}

\item{Fill}{setting Fill colour.}

\item{Size}{setting size of point, line,etc.}

\item{Alpha}{setting transparency.}

\item{Width}{setting width of barplot or boxplot.}

\item{Shape}{setting shape.}
}
\value{
resGGplot  ggplot.

resPlotly  plot that's generated with ggplotly.
}
\description{
use ggplot2S!
}
\examples{
ggplt2S(data=mtcars,x='mpg',y='disp',size='wt',geom=c('point','smooth'),smoothMethod='lm',theme='bw',title='my plot')

}
