% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmS.R
\name{glmS}
\alias{glmS}
\title{glmS}
\usage{
glmS(Formula, data, weightsVar = 1, subset = "all", Family = c("gaussian",
  "binomial", "poisson"), na.action = na.rm, lower = "~1")
}
\description{
is a function used in shiny app.
}
