% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hTest.R
\name{hTest}
\alias{hTest}
\title{hTest}
\usage{
hTest(data, xvars, yvars = "", alter = c("two.sided", "less", "greater")[1],
  paired = FALSE, confLevel = 0.95, nullHyp = 0, normalSampleSize = 100)
}
\arguments{
\item{data}{data.frame}

\item{xvars}{variables on the right hand side.}

\item{yvars}{variables on the left hand side. when yvars is null, statistical tests are made on xvars.}

\item{alter}{could be two.sided,less,greater.}

\item{paired}{logical, if the x variable and yvariable is paired.}

\item{confLevel}{confidence level,alpha=1-confLevel}

\item{nullHyp}{eg. mu=0 in t.test}

\item{normalSampleSize}{integer, when sample size exceed this value, always using parametric tests instead.}
}
\description{
hTest provides some basic hypothesis statistical tests, and can detect which test method should be used based on the variable
modes and distribution of continuous variables.
}
\details{
when there are multiple variables both in xvars and yvars params, hTest performs statistical tests on each x and y.
}
\examples{
hTest(iris,xvars=c('Sepal.Length','Sepal.Width'),yvars='Species',alter='two.sided')->res
res$hTestRes

plot(res$hTestGraph)


}
