% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{imputeData}
\usage{
imputeData(data, impVars, modelVars, seed = 1, treeParams = list(),
  method = c("sample", "pred")[1])
}
\description{
is a imputation function
}
