% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeS.R
\name{lmeS}
\alias{lmeS}
\title{lmeS}
\usage{
lmeS(formulaFixed, formulaRandom, Method = c("ML", "REML")[1], data,
  subset = "all")
}
\arguments{
\item{formulaFixed}{formula of fixed effects}

\item{formulaRandom}{formula of random effects}

\item{Method}{could be 'ML' or 'REML'.}

\item{data}{data.frame obj.}

\item{subset}{logical expression or 'all' which means no subsetting.}
}
\description{
lme function used in shiny app.
}
