% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Madis.R
\name{Madis}
\alias{Madis}
\title{Madis}
\usage{
Madis(wd = paste0(getwd(), "/"), lang = c("ch", "en")[1],
  Encod = "utf8", ...)
}
\arguments{
\item{wd}{set working directory.}

\item{lang}{en for english, cn for chinese version}

\item{...}{more args used in runApp, eg. launch.browser=T}
}
\description{
launcher function
}
