% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIt.R
\name{matchIt}
\alias{matchIt}
\title{matchIt}
\usage{
matchIt(formula, data, Method = "nearest", Distance = "logit",
  Ratio = 1)
}
\description{
matchIt is a simplyfied function based on matchIt from MatchIt package, and with some revisement used in madis
}
