% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaS.R
\name{pcaS}
\alias{pcaS}
\title{pcaS}
\usage{
pcaS(data, vars, nfcts = 2, Rotate = c("none", "varimax", "quartimax",
  "promax", "oblimin", "simplimax", "cluster")[1], Scores = T,
  subset = "all", pcaVarName = "", addVar = T)
}
\description{
PCA analysis used in shiny app.
}
