% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophetS.R
\name{prophetS}
\alias{prophetS}
\title{prophetS}
\usage{
prophetS(data, tsVar, tsFormat = "ymd", measureVars, groupVars = 1,
  Period = "weeks", FN, Cap = -1, Floor = -1, Growth = "linear",
  H = 10, yearlyS = "auto", dailyS = "auto", weeklyS = "auto")
}
\description{
is a revised version of prophet that combine some other functions from xts package for dealing data.frame, multiple variables and adding group variables. 
return the history data and graphs as well as predicted graph using prophet.
}
