% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickShinyFuns.R
\name{qGraph}
\alias{qGraph}
\title{qPlotly}
\usage{
qGraph(dt, ...)
}
\description{
quick plotly based on ggplt2S. along with Filter param.
}
\details{
use ggplt2S params in ...
}
\examples{

fnPlotly(dt=mtcars,x='disp',y='mpg',geom=c('point','smooth'),Colour='red',alpha=.5,Filter=c('vs','gear','am'))

}
