% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{recode}
\usage{
recode(x, groups, Labels = NULL, na.val = "others", method = c("ranges",
  "substrs", "elements")[1], mode = c("numeric", "character", "datetime")[1])
}
\arguments{
\item{x}{is a vector which could be numeric, datetime or character}

\item{groups}{is simlar with L in legal_set, but different in that for elements and substrs it should be a list}

\item{Labels}{defines the labels for the returned value}

\item{method}{could be ranges, substrs and elements}

\item{mode}{could be numeric, character or datetime}
}
\value{
a regrouped vector
}
\description{
recode is a function simlar with legal_set, but return a regrouped value from input x.
}
\examples{
x=rnorm(100)
recode(x,groups=c('(-Inf,0]','(0,Inf)'),Labels=c('negative_value','positive_value'),method='ranges',mode='numeric')
x=sample(letters[1:5],100,rep=T)
recode(x,groups=list(c('a','b'),c('c','d','e')),Labels=c('a-b','c-e'),method='elements',mode='character')
x=c('google','goodbye','doodle','doodoge','go!','bad')
recode(x,groups=list(c('goo'),c('doo')),Labels=c('include_goo','include_doo'),method='substrs',mode='character')

}
\author{
sontron
}
