% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{table1}
\usage{
table1(data, grpVars = "", testVars, Digits = 2, normSamSize = 30)
}
\arguments{
\item{data}{data.frame}

\item{grpVars}{grouping variable names in data. could also use ";" seperate different grouping variables, like grpVars="grp1;grp2"}

\item{testVars}{variable names in data. could also use ";" seperate different vars, like "age;sex"}
}
\description{
generate table1
}
\examples{

table1(iris,grpVars=c('Species'),testVars=c('Sepal.Length','Sepal.Width'))

table1(iris,grpVars=c('Species'),testVars=c('Sepal.Length;Sepal.Width'))



}
