% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeS.R
\name{treeS}
\alias{treeS}
\title{treeS}
\usage{
treeS(Formula, data, subset = "all", na.action = na.rm, Minsplit = 30,
  Minbucket = 10, Mincrit = 0.05, Maxdepth = 3, CP = 0.01,
  treeMethod = c("ctree", "rpart")[1])
}
\description{
is a function of Tree models
}
