% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniVarVec.R
\name{uniVarVec}
\alias{uniVarVec}
\title{uniVarVec}
\usage{
uniVarVec(x, varType = c("numeric", "character", "factor", "integer",
  "ordered")[1], Digits = 4, nameX = "x", seed = 123, tabSort = TRUE)
}
\description{
a function provide description statistics for vectors
}
