% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{validateData}
\alias{validateData}
\title{validateData}
\usage{
validateData(data, val, legal, method = c("ranges", "substrs",
  "elements")[1], mode = c("numeric", "character", "datetime")[1],
  type = c("fixed", "regex")[1], tsFormat = "ymd")
}
\arguments{
\item{data}{is a data.frame object}

\item{val}{is a character vector to specified the variables in data. It's same as x in validateVec}

\item{legal}{is a list to specified expression for each val. It's same as L in validateVec}

\item{method}{is a character vector specified methods for each variable in val param.}

\item{mode}{is a character vector specified modes for each variable in val.}

\item{tsFormat}{used to detect tsFormat in ts value}
}
\value{
logical value such as T,F,NA.
}
\description{
can be used for data.frame object
}
\examples{
validateData(data=iris,val=c('Species','Sepal.Length'),
legal=list(c('setosa','versicolor'),c('[3,6]')),method=c('elements','ranges'),
mode=c('character','numeric'))

}
\author{
sontron
}
