% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMnp.R
\name{dataMnp}
\alias{dataMnp}
\title{dataMnp}
\usage{
dataMnp(data, subset = NA, newVars = NA, newVarsFormulas = NA,
  newVarsBy = NA, newVarsMargin = NA, indexNames, Formulas, dimVars,
  dimNames = NA, dateVar = NA, dtOrders = "ymd", margin = 0,
  revisedMargin = 0, revisedNames = NA, revisedFormulas = NA,
  orderVars = NA, orders = NA, Digits = 2, tbVars = NA, hbVars = NA,
  colOrder = NA)
}
\description{
a function provides data manipulating procedures
}
\details{
add order param!!
}
\examples{
expand.grid(2017:2018,1:12,1:28)->X
apply(X,1,function(x)paste(x,collapse='-'))->date
dt<-data.frame(ts=date,grp1=sample(letters[1:5],length(date),rep=T),y=rnorm(length(date)))
dataMnp(data=dt,indexNames=c('a','b','c'),Formulas='mean(y);sd(y);sd(y)/mean(y)',dimVars='year(ts);month(ts);grp1',dimNames = 'year;month;grp',dateVar = 'ts',dtOrders = 'ymd',orderVars = 'grp',orders='1',tbVars='a;b',hbVars=c('c','tt'),revisedNames = 'tt',revisedFormulas = 'a+b',colOrder='1;2;3;4;5;6;8;9')->res




}
