% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateVec.R
\name{validateVec}
\alias{validateVec}
\title{validateVec}
\usage{
validateVec(x, L, method = c("ranges", "substrs", "elements")[1],
  mode = c("numeric", "character", "datetime")[1], type = c("fixed",
  "regex")[1])
}
\arguments{
\item{x}{is a vector which could be numeric, character or datetime type}

\item{L}{is a expression indicate legal sets for x}

\item{method}{specified the method should be used}

\item{mode}{is the mode of x}
}
\value{
logic value which could be T,F,NA
}
\description{
This function return logic value that indicate which x values are in legal set that is set by L
}
\note{
the expression of L is flexible,like "(0,3)","[7,11]","[0,Inf)",c('[-1,0]','(1,10)'),
"[2013-1-2,2-13-10-11]",c('apple','orange')
}
\examples{
x=rnorm(10)
L="[0,Inf]"
validateVec(x,L,method='ranges',mode='numeric')
x.char=c('a','b','cd')
L.char=c('a','c')
validateVec(x=x.char,L=L.char,method='elements',mode='character')
validateVec(x=x.char,L=L.char,method='substrs',mode='character')
Date=c('2016-1-1 14:24:00',NA,'2016-6-1','2016-12-31 24:00:00')
L.date='[2016-1-1,2016-12-31)'
validateVec(x=Date,L=L.date,method='ranges',mode='datetime')

}
\author{
sontron
}
